refactor(#3479): extract dispatch-policy cluster from discord/mod.rs to dispatch_policy.rs
Verbatim move of the 14-symbol dispatch-policy cluster (monitor-auto-turn origin
marker codec, stale-dispatch turn gating + its tests, allowed-turn-sender /
announce+notify bot-user-id resolvers, phase2 recovery predicate) to the new
capped sibling dispatch_policy.rs (181 prod LoC). discord/mod.rs production LoC
ratcheted 4243 -> 4074 (-169). No logic changes.
