+++ ARCHITECTURE.md
@@ -573,6 +573,7 @@ src/
 │   │   ├── answer_flush_barrier.rs
 │   │   ├── catch_up.rs
 │   │   ├── discord_io.rs
+│   │   ├── dispatch_policy.rs
 │   │   ├── formatting.rs
 │   │   ├── gateway.rs
 │   │   ├── health.rs
